+++ src/services/infrastructure/kratos/types/ory.traits.ts
@@ -1,0 +1,9 @@
+export interface OryTraits {
+  accepted_terms: boolean;
+  picture: string;
+  email: string;
+  name: {
+    first: string;
+    last: string;
+  };
+}
