+++ src/services/infrastructure/kratos/types/session.invalid.enum.ts
@@ -1,0 +1,4 @@
+export type SessionInvalidReason =
+  | 'Session not defined'
+  | 'Session expiry not defined'
+  | 'Session expired';
